% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mat.R
\name{norm_mat}
\alias{norm_mat}
\title{Normal matrix}
\usage{
norm_mat(n = 1, p = 1, mu = rep(0, p), sd = rep(1, p))
}
\arguments{
\item{n}{sample size}

\item{p}{number of variables}

\item{mu}{mean vector}

\item{sd}{diagonal of the covariance matrix}
}
\value{
normal matrix
}
\description{
Normal matrix
}
\examples{
norm_mat(1e2, 3, -1:1, 1:3)
}
